requirements.txt

tweepy
praw
newsapi-python
yfinance
transformers
scikit-learn
pandas
numpy
requests
python-dotenv


